+++ reikna/cluda/cuda.py
@@ -150,8 +150,10 @@ class Thread(api_base.Thread):
     def synchronize(self):
         self._queue.synchronize()
 
-    def _compile(self, src, fast_math=False):
+    def _compile(self, src, fast_math=False, compiler_options=None):
         options = ['-use_fast_math'] if fast_math else []
+        if compiler_options is not None:
+            options += compiler_options
         return SourceModule(src, no_extern_c=True, options=options)
 
     def _cuda_push(self):