+++ doc/source/history.rst
@@ -10,6 +10,8 @@ Release history
 
 * ADDED: a scan class :py:class:`~reikna.algorithms.Scan`.
 
+* ADDED: an optional parameter ``compiler_options`` for :py:meth:`Thread.compile <reikna.cluda.api.Thread.compile>`, :py:meth:`Thread.compile_static <reikna.cluda.api.Thread.compile_static>` and :py:meth:`Computation.compile <reikna.core.Computation.compile>`, allowing one to pass additional options to the compiler.
+
 
 0.6.8 (18 Dec 2016)
 ===================